+++ generators/csharp/codegen/src/ast/Argument.ts
@@ -1,0 +1,16 @@
+import { CodeBlock } from "./CodeBlock";
+
+export type Argument = NamedArgument | UnnamedArgument;
+
+export type Arguments = NamedArgument[] | UnnamedArgument[];
+
+export interface NamedArgument {
+    name: string;
+    assignment: CodeBlock;
+}
+
+export type UnnamedArgument = CodeBlock;
+
+export function isNamedArgument(argument: NamedArgument | UnnamedArgument): argument is NamedArgument {
+    return (argument as NamedArgument)?.name != null;
+}