+++ generators/csharp/sdk/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+- Internal: The C# generator now includes gRPC core utilities.
+
 ## [1.2.1 - 2024-08-12]
 
 - Feature: Add support for Protobuf file dependencies to generate gRPC client stubs.